+++ vault/client/src/components/Layout.jsx
@@ -333,7 +333,6 @@ function Layout() {
       <aside style={sidebarStyle}>
         <ProjectSidebar
           onClose={() => setSidebarOpen(false)}
-          showHabits={canUseFeature('goals') && canUseFeature('habitsSidebar')}
           showClientContext={/^\/projects\/\d+$/.test(location.pathname) || /^\/clients\/\d+/.test(location.pathname)}
           collapsed={!sidebarOpen && !isMobile}
         />